+++ apps/architect-vite/src/components/Routes.tsx
@@ -1,7 +1,7 @@
 import { Route, Switch } from "wouter";
 import Home from "~/components/Home/Home";
 import Protocol from "~/components/Protocol";
-import { AssetsPage, CodebookPage, StageEditorPage, SummaryPage } from "~/components/pages";
+import { AssetsPage, CodebookPage, ExperimentsPage, StageEditorPage, SummaryPage } from "~/components/pages";
 
 const Routes = () => {
 	return (
@@ -11,6 +11,7 @@ const Routes = () => {
 			<Route path="/protocol/codebook" component={CodebookPage} />
 			<Route path="/protocol/summary" component={SummaryPage} />
 			<Route path="/protocol/stage/:stageId" component={StageEditorPage} />
+			<Route path="/protocol/experiments" component={ExperimentsPage} />
 
 			<Route path="/" component={Home} />
 		</Switch>