+++ apps/architect-vite/src/components/Screens/NewStageScreen/InterfaceList.tsx
@@ -22,7 +22,7 @@ const InterfaceList = ({
 	<motion.div className="flex flex-col">
 		{items.length === 0 && (
 			<div className="absolute inset-0 flex flex-col items-center justify-center">
-				<p className="text-center mb-4">
+				<p className="text-center mb-4 w-4/5">
 					No interfaces match your filter and/or search results. Try a different combination of types, or clear your
 					filters and search query to see all available interfaces.
 				</p>