+++ apps/architect-vite/src/components/sections/NodePanels/NodePanels.tsx
@@ -71,7 +71,7 @@ const NodePanels = ({ form, createNewPanel, panels = null, disabled = false, ...
 
 				{!isFull && (
 					<div className="stage-editor-section-content-items__controls">
-						<Button onClick={() => createNewPanel()} icon="add">
+						<Button onClick={() => createNewPanel()} icon="add" color="sea-green">
 							Add new panel
 						</Button>
 					</div>