+++ apps/architect-vite/CHANGELOG.md
@@ -1,5 +1,25 @@
 # architect-vite
 
+## 7.1.0
+
+### Minor Changes
+
+- 01448c8: Split Family Tree sexVariable into egoSexVariable and nodeSexVariable.
+
+  This is a breaking change for existing protocols that reference the old sexVariable field. Protocols with Farmily Tree interfaces require that the egoSexVariable and nodeSexVariable be defined separately.
+
+### Patch Changes
+
+- Updated dependencies [01448c8]
+  - @codaco/protocol-validation@10.0.0
+
+## 7.0.3
+
+### Patch Changes
+
+- Updated dependencies [cc2adc3]
+  - @codaco/protocol-validation@9.0.0
+
 ## 7.0.2
 
 ### Patch Changes